// packages/cli.old/bin/interfaces.ts
interface FeatureSpec {
    id?: number;
    name: string;
    path: string;
    ext: FeatureExtension;
    variables?: TaskVariables | Record<string, any>;
    type?: string;
    category?: string;
}
// ...
type FeatureType = "task" | "agent" | "action" | "cmd" | "workflow" | "adaptater";
type ToolType = "task" | "agent" | "action" | "cmd" | "workflow";
type ActionExtension = "js" | "mjs" | "py" | "yml";
type TaskExtension = "yml";
type AgentExtension = "yml";
type AdaptaterExtension = "js";
type WorkflowExtension = "yml";
type CmdExtension = "js";
type FeatureExtension = TaskExtension | AgentExtension | CmdExtension | ActionExtension | WorkflowExtension;